fixed tests files
