+++ quiz-app/src/main/java/com/quizmaster/servlet/IndexServlet.java
@@ -1,0 +1,19 @@
+package com.quizmaster.servlet;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@WebServlet("/")
+public class IndexServlet extends HttpServlet {
+    
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) 
+            throws ServletException, IOException {
+        // Forward to index.jsp
+        request.getRequestDispatcher("/index.jsp").forward(request, response);
+    }
+}
