+++ quiz-app/src/main/webapp/profile.jsp
@@ -5,7 +5,89 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Profile - QuizMaster</title>
     <script src="https://cdn.tailwindcss.com"></script>
-    <link rel="stylesheet" href="profile.css">
+    <style>
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700&display=swap');
+
+body {
+    font-family: 'Inter', sans-serif;
+    background-color: #f2f0e9;
+    min-height: 100vh;
+}
+
+.card-hover {
+    transition: all 0.3s ease;
+}
+
+.card-hover:hover {
+    transform: translateY(-4px);
+    box-shadow: 0 12px 24px rgba(73, 89, 107, 0.15);
+}
+
+.sidebar-item {
+    transition: all 0.2s ease;
+}
+
+.sidebar-item:hover {
+    background-color: #f8fafc;
+    transform: translateX(2px);
+}
+
+.notification-pulse {
+    animation: pulse 2s infinite;
+}
+
+@keyframes pulse {
+    0%, 100% { opacity: 1; }
+    50% { opacity: 0.7; }
+}
+
+.profile-picture {
+    transition: all 0.3s ease;
+    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
+}
+
+.profile-picture:hover {
+    transform: scale(1.05);
+    box-shadow: 0 12px 40px rgba(0, 0, 0, 0.15);
+}
+
+/* Stats Cards */
+.stat-card-1 {
+    background: linear-gradient(135deg, #8387C3 0%, #676a9c 100%) !important;
+    color: #000;
+    position: relative;
+    overflow: hidden;
+    border: 2px solid #222;
+    z-index: 1;
+}
+
+.stat-card-2 {
+    background: linear-gradient(135deg, #CEB5D4 0%, #a68eab 100%) !important;
+    color: #000;
+    position: relative;
+    overflow: hidden;
+    border: 2px solid #222;
+    z-index: 1;
+}
+
+.stat-card-3 {
+    background: linear-gradient(135deg, #537E72 0%, #406358 100%) !important;
+    color: #000;
+    position: relative;
+    overflow: hidden;
+    border: 2px solid #222;
+    z-index: 1;
+}
+
+.stat-card-4 {
+    background: linear-gradient(135deg, #7D9FC0 0%, #5f7b96 100%) !important;
+    color: #000;
+    position: relative;
+    overflow: hidden;
+    border: 2px solid #222;
+    z-index: 1;
+}
+    </style>
 </head>
 <body class="overflow-hidden bg-gray-50">
 <div class="flex h-screen">
