+++ quiz-app/src/main/webapp/quizSummary.jsp
@@ -28,7 +28,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Quiz Summary - <%= quiz.getTitle() %></title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
+    <link rel="stylesheet" href="<c:url value='/styles.css'/>">
     <style>
         .quiz-header {
             background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
