+++ quiz-app/src/test/java/models/QuizResultTest.java
@@ -1,5 +1,6 @@
 package models;
 
+import models.QuizResult;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
