+++ quiz-app/src/test/java/notifications/ChallengeNotificationTest.java
@@ -1,6 +1,7 @@
 package notifications;
 
 import models.Notification;
+import notifications.ChallengeNotification;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
