+++ quiz-app/src/test/java/notifications/FriendRequestNotificationTest.java
@@ -1,54 +1,55 @@
 package notifications;
 
 import models.Notification;
+import notifications.FriendRequestNotification;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeEach;
 import static org.junit.jupiter.api.Assertions.*;
 
-public class NoteNotificationTest {
+public class FriendRequestNotificationTest {
 
-    private NoteNotification notification;
+    private FriendRequestNotification notification;
 
     @BeforeEach
     void setUp() {
-        notification = new NoteNotification();
+        notification = new FriendRequestNotification();
     }
 
     @Test
     void testDefaultConstructor() {
-        NoteNotification notification = new NoteNotification();
+        FriendRequestNotification notification = new FriendRequestNotification();
 
         assertEquals(0, notification.getId());
         assertEquals(0, notification.getFromUserId());
         assertEquals(0, notification.getToUserId());
         assertNull(notification.getTitle());
         assertNull(notification.getMessage());
-        assertEquals(Notification.NOTE_NOTIFICATION, notification.getQuestionType());
+        assertEquals(Notification.FRIEND_REQUEST_NOTIFICATION, notification.getQuestionType());
     }
 
     @Test
     void testParameterizedConstructor() {
         long fromUserId = 123L;
         long toUserId = 456L;
-        String title = "Note Title";
-        String message = "Note Message";
+        String title = "Friend Request Title";
+        String message = "Friend Request Message";
 
-        NoteNotification notification = new NoteNotification(fromUserId, toUserId, title, message);
+        FriendRequestNotification notification = new FriendRequestNotification(fromUserId, toUserId, title, message);
 
         assertEquals(0, notification.getId()); // ID is set to 0 in parent constructor
         assertEquals(fromUserId, notification.getFromUserId());
         assertEquals(toUserId, notification.getToUserId());
         assertEquals(title, notification.getTitle());
         assertEquals(message, notification.getMessage());
-        assertEquals(Notification.NOTE_NOTIFICATION, notification.getQuestionType());
+        assertEquals(Notification.FRIEND_REQUEST_NOTIFICATION, notification.getQuestionType());
     }
 
     @Test
     void testSettersAndGetters() {
         long fromUserId = 789L;
         long toUserId = 101112L;
-        String title = "New Note";
-        String message = "New Note Message";
+        String title = "New Friend Request";
+        String message = "New Friend Request Message";
         long id = 999L;
 
         notification.setFromUserId(fromUserId);
@@ -62,12 +63,12 @@ public class NoteNotificationTest {
         assertEquals(title, notification.getTitle());
         assertEquals(message, notification.getMessage());
         assertEquals(id, notification.getId());
-        assertEquals(Notification.NOTE_NOTIFICATION, notification.getQuestionType());
+        assertEquals(Notification.FRIEND_REQUEST_NOTIFICATION, notification.getQuestionType());
     }
 
     @Test
     void testQuestionTypeConstant() {
-        assertEquals("NOTE_NOTIFICATION", Notification.NOTE_NOTIFICATION);
+        assertEquals("FRIEND_REQUEST_NOTIFICATION", Notification.FRIEND_REQUEST_NOTIFICATION);
     }
 
     @Test
@@ -79,8 +80,8 @@ public class NoteNotificationTest {
         assertEquals("DIFFERENT_TYPE", notification.getQuestionType());
 
         // Create new instance to verify constructor behavior
-        NoteNotification newNotification = new NoteNotification(1L, 2L, "title", "message");
-        assertEquals(Notification.NOTE_NOTIFICATION, newNotification.getQuestionType());
+        FriendRequestNotification newNotification = new FriendRequestNotification(1L, 2L, "title", "message");
+        assertEquals(Notification.FRIEND_REQUEST_NOTIFICATION, newNotification.getQuestionType());
     }
 
     @Test
@@ -90,50 +91,38 @@ public class NoteNotificationTest {
 
     @Test
     void testWithNullValues() {
-        NoteNotification notification = new NoteNotification(0L, 0L, null, null);
+        FriendRequestNotification notification = new FriendRequestNotification(0L, 0L, null, null);
 
         assertEquals(0L, notification.getFromUserId());
         assertEquals(0L, notification.getToUserId());
         assertNull(notification.getTitle());
         assertNull(notification.getMessage());
-        assertEquals(Notification.NOTE_NOTIFICATION, notification.getQuestionType());
+        assertEquals(Notification.FRIEND_REQUEST_NOTIFICATION, notification.getQuestionType());
     }
 
     @Test
     void testWithEmptyStrings() {
-        NoteNotification notification = new NoteNotification(1L, 2L, "", "");
+        FriendRequestNotification notification = new FriendRequestNotification(1L, 2L, "", "");
 
         assertEquals("", notification.getTitle());
         assertEquals("", notification.getMessage());
-        assertEquals(Notification.NOTE_NOTIFICATION, notification.getQuestionType());
+        assertEquals(Notification.FRIEND_REQUEST_NOTIFICATION, notification.getQuestionType());
     }
 
     @Test
-    void testNoteSpecificScenarios() {
-        // Test typical note scenario
-        long senderId = 300L;
-        long receiverId = 400L;
-        String title = "Reminder";
-        String message = "Don't forget about the meeting tomorrow";
+    void testFriendRequestSpecificScenarios() {
+        // Test typical friend request scenario
+        long senderId = 100L;
+        long receiverId = 200L;
+        String title = "Friend Request";
+        String message = "John Doe wants to be your friend";
 
-        NoteNotification note = new NoteNotification(senderId, receiverId, title, message);
+        FriendRequestNotification friendRequest = new FriendRequestNotification(senderId, receiverId, title, message);
 
-        assertEquals(senderId, note.getFromUserId());
-        assertEquals(receiverId, note.getToUserId());
-        assertEquals(title, note.getTitle());
-        assertEquals(message, note.getMessage());
-        assertEquals(Notification.NOTE_NOTIFICATION, note.getQuestionType());
-    }
-
-    @Test
-    void testLongMessageContent() {
-        String longMessage = "This is a very long note message that contains multiple sentences. " +
-                "It might be used for detailed instructions or important information. " +
-                "The system should handle long messages without any issues.";
-
-        NoteNotification note = new NoteNotification(1L, 2L, "Long Note", longMessage);
-
-        assertEquals(longMessage, note.getMessage());
-        assertEquals(Notification.NOTE_NOTIFICATION, note.getQuestionType());
+        assertEquals(senderId, friendRequest.getFromUserId());
+        assertEquals(receiverId, friendRequest.getToUserId());
+        assertEquals(title, friendRequest.getTitle());
+        assertEquals(message, friendRequest.getMessage());
+        assertEquals(Notification.FRIEND_REQUEST_NOTIFICATION, friendRequest.getQuestionType());
     }
 }
