+++ quiz-app/src/test/java/users/UserTest.java
@@ -1,3 +1,5 @@
+package users;
+
 import models.User;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeEach;
